+++ src/main.ts
@@ -5,10 +5,13 @@ import { lockPullRequest } from './pullrequest/pullRequestLock'
 import * as core from '@actions/core'
 import * as input from './shared/getInputs'
 import { validateMergedPullRequestForLock } from './livePullRequest'
+import { requireHttpsDocumentUrl } from './shared/documentUrl'
 
 export async function run() {
   try {
     core.info(`CLA Assistant GitHub Action bot has started the process`)
+
+    requireHttpsDocumentUrl()
 
     if (!input.getRequiredBaseRef()) {
       core.warning(